Split addEvent/s into different files for server and browser, clean up client.get for browser
