+++ src/server/lib/addEvent.js
