+++ test/browser/examples/query/index.html
@@ -12,7 +12,7 @@
 			readKey: "4e4d72e5bf8b69686ed87a5a9671bba7ad829fbd10a1c281ee51b6e9c1ce9548e941e1f336f9de9281a5acc66ca8fdabc9b3c806e390eca01665f6a308a9b03d8b332b3fbd9f3cdfc3b3e16b0da6d84851e53fe20fbbce300801b8a401a6395b9f4ab9c89bff566e9678a74ca6624f9b"
 			// , protocol: "https"
 			// , host: 'analytics.yourdomain.com'
-			, requestType: "jsonp" // "jsonp" "beacon"
+			, requestType: "beacon" // "jsonp" "beacon"
 		});
 
 		/*
